Refactor MapDB and InMemory StatsStorage; add BaseCollectionStatsStorage; make InMemoryStatsStorage not rely on MapDB (for j7 compat)
