+++ deeplearning4j-ui-parent/deeplearning4j-ui-model/src/main/java/org/deeplearning4j/ui/storage/InMemoryStatsStorage.java
@@ -1,9 +1,17 @@
 package org.deeplearning4j.ui.storage;
 
+import org.deeplearning4j.api.storage.Persistable;
+import org.deeplearning4j.api.storage.StatsStorageEvent;
+import org.deeplearning4j.api.storage.StatsStorageListener;
+import org.deeplearning4j.api.storage.StorageMetaData;
 import org.deeplearning4j.ui.storage.mapdb.MapDBStatsStorage;
 import org.deeplearning4j.util.UIDProvider;
 
+import java.io.IOException;
+import java.util.List;
+import java.util.Map;
 import java.util.UUID;
+import java.util.concurrent.ConcurrentHashMap;
 
 /**
  * A StatsStorage implementation that stores all data in memory. If persistence is required for the UI information,
@@ -12,14 +20,83 @@ import java.util.UUID;
  *
  * @author Alex Black
  */
-public class InMemoryStatsStorage extends MapDBStatsStorage {
+public class InMemoryStatsStorage extends BaseCollectionStatsStorage {
     private final String uid;
 
     public InMemoryStatsStorage(){
         super();
         String str = UUID.randomUUID().toString();
         uid = str.substring(0,Math.min(str.length(),8));
+
+        
     }
+
+
+    @Override
+    protected synchronized Map<Long, Persistable> getUpdateMap(String sessionID, String typeID, String workerID, boolean createIfRequired) {
+        SessionTypeWorkerId id = new SessionTypeWorkerId(sessionID, typeID, workerID);
+        if (updates.containsKey(id)) {
+            return updates.get(id);
+        }
+        if (!createIfRequired) {
+            return null;
+        }
+        Map<Long, Persistable> updateMap = new ConcurrentHashMap<>();
+        updates.put(id, updateMap);
+        return updateMap;
+    }
+
+    @Override
+    public void putStaticInfo(Persistable staticInfo) {
+        List<StatsStorageEvent> sses = checkStorageEvents(staticInfo);
+        if(!sessionIDs.contains(staticInfo.getSessionID())){
+            sessionIDs.add(staticInfo.getSessionID());
+        }
+        SessionTypeWorkerId id = new SessionTypeWorkerId(staticInfo.getSessionID(), staticInfo.getTypeID(), staticInfo.getWorkerID());
+
+        this.staticInfo.put(id, staticInfo);
+        StatsStorageEvent sse = null;
+        if (listeners.size() > 0) sse = new StatsStorageEvent(this, StatsStorageListener.EventType.PostStaticInfo,
+                staticInfo.getSessionID(), staticInfo.getTypeID(), staticInfo.getWorkerID(), staticInfo.getTimeStamp());
+        for (StatsStorageListener l : listeners) {
+            l.notify(sse);
+        }
+
+        notifyListeners(sses);
+    }
+
+    @Override
+    public void putUpdate(Persistable update) {
+
+    }
+
+    @Override
+    public void putStorageMetaData(StorageMetaData storageMetaData) {
+        List<StatsStorageEvent> sses = checkStorageEvents(storageMetaData);
+        SessionTypeId id = new SessionTypeId(storageMetaData.getSessionID(), storageMetaData.getTypeID());
+        this.storageMetaData.put(id, storageMetaData);
+
+        StatsStorageEvent sse = null;
+        if (listeners.size() > 0) sse = new StatsStorageEvent(this, StatsStorageListener.EventType.PostMetaData,
+                storageMetaData.getSessionID(), storageMetaData.getTypeID(), storageMetaData.getWorkerID(), storageMetaData.getTimeStamp());
+        for (StatsStorageListener l : listeners) {
+            l.notify(sse);
+        }
+
+        notifyListeners(sses);
+    }
+
+
+    @Override
+    public void close() throws IOException {
+        //No op
+    }
+
+    @Override
+    public boolean isClosed() {
+        return false;
+    }
+
 
     @Override
     public String toString(){
